Change the send_message trigger time
With the current method through the tornado_apns project, we send message to
the APNS immediatly when the connect done. And we do a connect-disconnect
circle at every time we send a message. But the apple's document said "APNs
treats rapid connection and disconnection as a denial-of-service attack."
So, when we have large messages to send, we'd better keep a connectiong alive
for a long peroid.
I use the tornado_apns in my project like below:
Create a APNS instance
connect to the server for inition
Use send_notification when we have a send request